Created PreresultTop and fixed the positioning
