+++ Comps/Kozy/CustomButton/custombutton.css
@@ -2,6 +2,8 @@
     display: flex;
     justify-content: center;
     align-items: center;
+    position: relative;
+    top: -20pt;
 
 }
